fix: place footer below navbar
