+++ sunny_sales_web/src/pages/StatsScreen.jsx
@@ -3,12 +3,11 @@ import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { BASE_URL } from '../config';
 import { BarChart, Bar, XAxis, YAxis, Tooltip, ResponsiveContainer, CartesianGrid } from 'recharts';
-import { useNavigate } from 'react-router-dom';
+import BackHomeButton from '../components/BackHomeButton';
 
 // Mostra gráfico de distâncias percorridas pelo vendedor
 export default function StatsScreen() {
   const [chartData, setChartData] = useState([]);
-  const navigate = useNavigate();
 
   // Carrega os trajetos para gerar o gráfico
   const loadRoutes = async () => {
@@ -47,7 +46,7 @@ export default function StatsScreen() {
 
   return (
     <div style={styles.container}>
-      <button onClick={() => navigate(-1)} style={styles.back}>⬅ Voltar</button>
+      <BackHomeButton />
       <h2 style={styles.title}>Distâncias percorridas por dia</h2>
 
       {chartData.length > 0 ? (
@@ -73,12 +72,4 @@ export default function StatsScreen() {
 const styles = {
   container: { padding: '2rem', maxWidth: 800, margin: '0 auto' },
   title: { fontSize: '1.5rem', marginBottom: '1rem' },
-  back: {
-    marginBottom: '1rem',
-    background: 'none',
-    border: 'none',
-    cursor: 'pointer',
-    fontSize: '1rem',
-    color: '#000',
-  },
 };
